--- ndspflow/reports/templates/masthead.html
@@ -1,39 +1,0 @@
-<html>
-<head>
-  <script src="{% ROOT %}/resources/js/plotly-latest.min.js"></script>
-  <script src="{% ROOT %}/resources/js/navbar.js"></script>
-  <script src="{% OUT %}/bycycle/.data.js" ></script>
-  <link rel="stylesheet" href="{% ROOT %}/resources/css/style.css">
-</head>
-<body>
-<div class="masthead">
-  <div class="masthead__inner-wrap">
-    <div class="masthead__menu">
-      <div class="site-title-box">
-        <a class="site-title">ndsp<a class="site-title__sub">flow</a></a>
-      </div>
-      <header id="navbar">
-        <nav class="navbar-container container">
-          <button type="button" class="navbar-toggle" aria-label="Open navigation menu">
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-          </button>
-          <div class="navbar-menu">
-          <ul class="navbar-links">
-            <li class="navbar-item"><a class="navbar-link" href="{% NAV_FOOOF %}">{% FOOOF %}</a></li>
-            <li class="navbar-item"><a class="navbar-link" href="{% NAV_BYCYCLE %}">{% Bycycle %}</a></li>
-          </ul>
-          </div>
-        </nav>
-      </header>
-    </div>
-  </div>
-</div>
-{% BODY %}
-</body>
-</html>
-
-<script>
-  makeNavbar()
-</script>
