--- ndspflow/reports/templates/nav.html
@@ -1,9 +1,0 @@
-<div class="container-body">
-  <div class="sub-container">
-    <h1>Group Info</h1>
-    {% META_TEMPLATE %}
-  </div>
-  <div class="sub-container">
-    {% NAV %}
-  </div>
-</div>
