--- ndspflow/reports/templates/resources/css/style.css
@@ -1,321 +1,0 @@
-html {
-  position: relative;
-  min-height: 100%;
-}
-body {
-  margin: 0;
-  color: #555555;
-  font-family: 'Roboto Condensed', sans-serif;
-  max-width: 1500px;
-  margin-left: auto;
-  margin-right: auto;
-}
-h1{
-  font-size: 1.5rem;
-  color: #555555;
-  font-family: 'Roboto Condensed', sans-serif;
-  font-weight: bold;
-}
-* {
-  box-sizing: border-box;
-}
-.masthead{
-  position: fixed;
-  width: 100%;
-  max-width: 1500px;
-  top: 0;
-  line-height: inherit;
-  background-color: #fff;
-  border-bottom: 1px solid #f2f3f3;
-  padding-left: 5em;
-  padding-right: 5em;
-  z-index: 20;
-}
-.masthead__menu {
-  display: flex;
-  margin-left: 0;
-  margin-right: 0;
-  width: 100%;
-  clear: both;
-  height: 5em;
-}
-.masthead__menu ul {
-  margin: 0;
-  padding: 0;
-  clear: both;
-  list-style-type: none;
-}
-.masthead__menu-item {
-  display: inline;
-  margin-left: 0.25em;
-  margin-right: 0.25em;
-  font-size: 1.5rem;
-  color: #555555;
-  font-family: 'Roboto Condensed', sans-serif;;
-  font-weight: lighter;
-
-  cursor: pointer;
-  vertical-align: middle;
-}
-.site-title-box{
-  margin-left: 0;
-  margin-top: auto;
-  margin-bottom: auto;
-  padding-right: 3em;
-}
-.site-title-box:hover .site-title{
-  color:  mix(#000, #555555, 10%);
-}
-.site-title-box:hover .site-title__sub{
-  color:  mix(#000, #de8435, 10%);
-}
-.site-title{
-  font-size: 2rem;
-  color: #555555;
-  font-family: 'Roboto Condensed', sans-serif;;
-  font-weight: bold;
-
-}
-.site-title__sub{
-  font-size: 1.5rem;
-  color: #de8435;
-  font-family: 'Roboto Condensed', sans-serif;;
-  font-weight: bold;
-}
-
-#navbar {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  flex: 1;
-  overflow: hidden;
-  margin-right: 2em;
-}
-.navbar-container {
-  display: flex;
-  justify-content: space-between;
-  height: 100%;
-  align-items: center;
-  left: 0;
-}
-.nav-btn{
-  position: relative;
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-}
-.navbar-item {
-  margin: 0.4em;
-  width: 100%;
-}
-.navbar-link {
-  font-size: 1.4rem;
-  color: #555555;
-  font-family: 'Roboto Condensed', sans-serif;;
-  font-weight: lighter;
-
-  transition: color 0.2s ease-in-out;
-  text-decoration: none;
-  display: flex;
-  font-weight: 400;
-  align-items: center;
-  transition: background-color 0.2s ease-in-out,
-    color 0.2s ease-in-out;
-  margin: 0 1rem;
-}
-.navbar-link:focus{
-  color:  mix(#000, #555555, 50%);
-  background-color: var(--navbar-bg-contrast);
-}
-.navbar-link:hover {
-  color:  mix(#000, #555555, 50%);
-  background-color: var(--navbar-bg-contrast);
-}
-.navbar-logo {
-  background-color: mix(#000, #555555, 50%);;
-  border-radius: 50%;
-  width: 30px;
-  height: 30px;
-  margin-right: 0.5em;
-}
-.navbar-toggle {
-  cursor: pointer;
-  border: none;
-  background-color: transparent;
-  width: 40px;
-  height: 40px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-.icon-bar {
-  display: block;
-  width: 25px;
-  height: 4px;
-  margin: 2px;
-  transition: background-color 0.2s ease-in-out,
-    transform 0.2s ease-in-out,
-    opacity 0.2s ease-in-out;
-  background-color: #555555;
-}
-.navbar-toggle:focus .icon-bar{
-  background-color:  mix(#000, #555555, 50%);
-}
-.navbar-toggle:hover .icon-bar {
-  background-color:  mix(#000, #555555, 50%);
-}
-#navbar.opened .navbar-toggle .icon-bar:first-child{
-  position: absolute;
-  margin: 0;
-  width: 30px;
-}
-#navbar.opened .navbar-toggle .icon-bar:last-child {
-  position: absolute;
-  margin: 0;
-  width: 30px;
-}
-#navbar.opened .navbar-toggle .icon-bar:first-child {
-  transform: rotate(45deg);
-}
-#navbar.opened .navbar-toggle .icon-bar:nth-child(2) {
-  opacity: 0;
-}
-#navbar.opened .navbar-toggle .icon-bar:last-child {
-  transform: rotate(-45deg);
-}
-.navbar-menu {
-  display: inline-flex;
-  position: fixed;
-  transition: opacity 0.2s ease-in-out,
-    visibility 0.2s ease-in-out;
-  opacity: 0;
-  visibility: hidden;
-  right:0;
-  top:4em;
-}
-#navbar.opened .navbar-menu {
-  background-color: #fff;
-  border-color: #555555;
-  opacity: 1;
-  visibility: visible;
-  position: right;
-}
-.navbar-links {
-  overflow: hidden;
-  position: top;
-  background-color: var(--navbar-bg-color);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  left: 0;
-  right: 0;
-  margin: 1.4rem;
-  border-radius: 5px;
-  box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
-}
-#navbar.opened .navbar-links {
-  padding: 1em;
-  max-height: none;
-}
-@media screen and (min-width: 880px) {
-  .navbar-toggle {
-  display: none;
-  }
-
-  #navbar .navbar-menu{
-  visibility: visible;
-  opacity: 1;
-  position: static;
-  display: block;
-  height: 100%;
-  }
-  #navbar.opened .navbar-menu {
-  visibility: visible;
-  opacity: 1;
-  position: static;
-  display: block;
-  height: 100%;
-  }
-
-  #navbar .navbar-links{
-  margin: 0;
-  padding: 0;
-  box-shadow: none;
-  position: static;
-  flex-direction: row;
-  list-style-type: none;
-  max-height: max-content;
-  width: 100%;
-  height: 100%;
-  }
-  #navbar.opened .navbar-links {
-  margin: 0;
-  padding: 0;
-  box-shadow: none;
-  position: static;
-  flex-direction: row;
-  list-style-type: none;
-  max-height: max-content;
-  width: 100%;
-  height: 100%;
-  }
-
-  #navbar .navbar-link:last-child {
-  margin-right: 0;
-  }
-}
-.container-body{
-  margin-top: 6rem;
-  padding-left: 5em;
-  padding-right: 5em;
-  font-size: 1rem;
-  color: #555555;
-  font-family: 'Roboto Condensed', sans-serif;;
-  font-weight: lighter;
-}
-.sub-container{
-  border-bottom: 1px solid #f2f3f3;
-  margin-bottom: 2em;
-}
-.plot-row {
-display: flex;
-  flex-wrap: wrap;
-}
-.plot-column {
-  flex: 0 0 50%;
-  padding: 10px;
-}
-.collapsible {
-  background-color: #ffffff;
-  cursor: pointer;
-  width: 100%;
-  border: none;
-  text-align: left;
-  outline: none;
-}
-.collapsible:hover {
-  background-color: #f2f2f2;
-}
-.content {
-  display: none;
-  overflow: hidden;
-}
-.btn {
-  display: inline-block;
-  margin-bottom: 0.25em;
-  padding: 0.5em 1em;
-  font-family: 'Roboto Condensed', sans-serif;
-  font-weight: bold;
-  text-align: center;
-  text-decoration: none;
-  border-width: 0;
-  cursor: pointer;
-  min-width:12em;
-  color: #555555;
-  border: 2px solid #555555;
-}
-
-a, a:visited, a:hover, a:active {
-  color: inherit;
-}
