--- ndspflow/reports/templates/resources/js/navbar.js
@@ -1,46 +1,0 @@
-function makeNavbar(){
-  const navbar = document.getElementById("navbar");
-  const navbarToggle = navbar.querySelector(".navbar-toggle");
-
-  function openMobileNavbar() {
-    navbar.classList.add("opened");
-    navbarToggle.setAttribute("aria-label", "Close navigation menu.");
-  }
-
-  function closeMobileNavbar() {
-    navbar.classList.remove("opened");
-    navbarToggle.setAttribute("aria-label", "Open navigation menu.");
-  }
-
-  navbarToggle.addEventListener("click", () => {
-    if (navbar.classList.contains("opened")) {
-      closeMobileNavbar();
-    } else {
-      openMobileNavbar();
-    }
-  });
-
-  const navbarMenu = navbar.querySelector(".navbar-menu");
-  const navbarLinksContainer = navbar.querySelector(".navbar-links");
-
-  navbarLinksContainer.addEventListener("click", (clickEvent) => {
-    clickEvent.stopPropagation();
-  });
-
-  navbarMenu.addEventListener("click", closeMobileNavbar);
-
-  var coll = document.getElementsByClassName("collapsible");
-  var i;
-  for (i = 0; i < coll.length; i++) {
-    coll[i].addEventListener("click", function() {
-      this.classList.toggle("active");
-      var content = this.nextElementSibling;
-      if (content.style.display === "block") {
-        content.style.display = "none";
-      } else {
-        content.style.display = "block";
-      }
-    });
-  }
-}
-
