--- ndspflow/reports/templates/subject.html
@@ -1,12 +1,0 @@
-<div class="container-body">
-  <div class="sub-container">
-    <h1>Individual Info</h1>
-    {% META_TEMPLATE %}
-  </div>
-  <div class="sub-container" id="fooof-container">
-    <h1>{% model_type %}</h1>
-    <p><center>{% settings %}</center></p>
-    <p><center>{% results %}</center></p>
-    <p><center>{% graph %}</center></p>
-  </div>
-</div>
